backend/requirements.txt
fastapi
uvicorn[standard]
chromadb
pymupdf
python-docx
beautifulsoup4
rank-bm25
httpx
python-dotenv
python-multipart
